LottoPrizeCalculator.kt 보완 (enum class 사용)

+++ src/main/kotlin/lotto/LottoRateOfReturn.kt
@@ -6,7 +6,7 @@ class LottoRateOfReturn {
     companion object {
         private const val ONE_HUNDRED = 100
     }
-    fun calculateRateOfReturn(totalPrize: Int, amount: Int): BigDecimal {
+    fun calculateRateOfReturn(totalPrize: BigDecimal, amount: Int): BigDecimal {
         return (totalPrize.toDouble() / amount * ONE_HUNDRED).toBigDecimal()
     }
 }
